WIP - RemovedLocation loop and moved classes around
