+++ src/main/java/com/hindsight/king_of_castrop_rauxel/action/ExitAction.java
@@ -17,7 +17,7 @@ public class ExitAction implements Action {
 
   @Setter private int index;
   @Setter private String name;
-  private static final State NEXT_STATE = State.AT_DEFAULT_POI;
+  private static final State NEXT_STATE = State.AT_POI;
 
   @Override
   public void execute(Player player) {
