+++ src/main/java/com/hindsight/king_of_castrop_rauxel/configuration/AppConfiguration.java
@@ -3,8 +3,8 @@ package com.hindsight.king_of_castrop_rauxel.configuration;
 import com.hindsight.king_of_castrop_rauxel.graphs.Graph;
 import com.hindsight.king_of_castrop_rauxel.location.AbstractLocation;
 import com.hindsight.king_of_castrop_rauxel.utils.*;
-import com.hindsight.king_of_castrop_rauxel.world.BasicTerrainGenerator;
-import com.hindsight.king_of_castrop_rauxel.world.TerrainGenerator;
+import com.hindsight.king_of_castrop_rauxel.utils.BasicTerrainGenerator;
+import com.hindsight.king_of_castrop_rauxel.utils.TerrainGenerator;
 import com.hindsight.king_of_castrop_rauxel.world.World;
 import com.hindsight.king_of_castrop_rauxel.world.WorldHandler;
 import java.util.Scanner;
