+++ src/main/java/com/hindsight/king_of_castrop_rauxel/encounter/EncounterSequence.java
@@ -1,10 +1,10 @@
-package com.hindsight.king_of_castrop_rauxel.cli.combat;
+package com.hindsight.king_of_castrop_rauxel.encounter;
 
 import com.hindsight.king_of_castrop_rauxel.characters.BasicEnemy;
 import com.hindsight.king_of_castrop_rauxel.characters.Combatant;
 import com.hindsight.king_of_castrop_rauxel.characters.Player;
+import com.hindsight.king_of_castrop_rauxel.cli.combat.Encounter;
 import com.hindsight.king_of_castrop_rauxel.event.Event;
-import com.hindsight.king_of_castrop_rauxel.location.DungeonDetails;
 import com.hindsight.king_of_castrop_rauxel.utils.Generators;
 import java.util.ArrayList;
 import java.util.List;
