+++ src/main/java/com/hindsight/king_of_castrop_rauxel/utils/NameGenerator.java
@@ -1,6 +1,7 @@
 package com.hindsight.king_of_castrop_rauxel.utils;
 
 import com.hindsight.king_of_castrop_rauxel.characters.Npc;
+import com.hindsight.king_of_castrop_rauxel.encounter.DungeonDetails;
 import com.hindsight.king_of_castrop_rauxel.location.*;
 
 public interface NameGenerator extends Generator {
